Run multiple jobs per fork (#52)
* Run multiple jobs per fork

By default Resque forks a new process for each job it processes. This approach
is not well suited for us, as we do lots of tiny jobs and the overhead of
forking a process is actually larger than the real execution time. With this
patch 50 jobs are run in a single fork, leading to much better performance.

Before this a single executor could deliver 1000 messages in around 6 minutes
and 42 seconds. With this patch and 50 jobs per fork, 1000 messages can be
delivered in roughly 45 seconds. Further tweaking may improve this more, 50 was
just a guess.

* Allow producing many messages at once

* Make rubocop happy

* Add benchmark script

* Workaround producer's buffer size

* Right align table

* Address comments